% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChiSquareTest.R
\name{ChiSquareTest}
\alias{ChiSquareTest}
\title{Chi-Square test}
\usage{
ChiSquareTest(JT, v1, v2)
}
\arguments{
\item{JT}{- joint distribution table used}

\item{v1}{- first variable}

\item{v2}{- second variable}
}
\value{
chi_square - chi_square test statistic

p_value - p value in the test
}
\description{
Test the independence of two variables
}

