% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification.R
\name{Classification}
\alias{Classification}
\title{Classification}
\usage{
Classification(data, theta, psi)
}
\arguments{
\item{data}{- Imputed dataset}

\item{theta}{- vector of probability for each component}

\item{psi}{- specific probability for each variable in each component}
}
\value{
class - class for each observation.
}
\description{
This function is used to classify each observation.
}

