% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteData.R
\docType{data}
\name{CompleteData}
\alias{CompleteData}
\title{This is an example for complete dataset}
\description{
CompleteData dataset is an array with parameters n = 10, p = 2, d = 2, m=10, k = 2.
}
\author{
Chaojie Wang \email{wang910930@163.com}
}
\keyword{data}

