% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation.R
\name{Correlation}
\alias{Correlation}
\title{Correlation estimation}
\usage{
Correlation(JT)
}
\arguments{
\item{JT}{- joint distribution table used}
}
\value{
cor - correlation matrix
}
\description{
Calculate correlation matrix among variables
}

