% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBhattacharyya.R
\name{DBhattacharyya}
\alias{DBhattacharyya}
\title{Bhattacharyya distance}
\usage{
DBhattacharyya(JT1, JT2)
}
\arguments{
\item{JT1}{- first joint distribution table}

\item{JT2}{- second joint distribution table}
}
\description{
Bhattacharyya distance
}

