% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example.R
\docType{data}
\name{Example}
\alias{Example}
\title{Real application dataset}
\description{
This is a real application dataset.
The source of original data is the ratings dataset in (Harper and Konstan (2016) <DOI:10.1145/2827872>).
This dataset is used to evaluate the performance of package in real applications.
}
\author{
Chaojie Wang
}
\keyword{data}

