% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorToMatrix.R
\name{FactorToMatrix}
\alias{FactorToMatrix}
\title{Factor To Matrix}
\usage{
FactorToMatrix(data_f, d)
}
\arguments{
\item{data_f}{- data in factor formation}

\item{d}{- number of choices in each variable}
}
\value{
data_m - data in matrix formation
}
\description{
This function is used to transform data from factor into matrix.
}

