% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FisherTest.R
\name{FisherTest}
\alias{FisherTest}
\title{Fisher exact test}
\usage{
FisherTest(theta, psi, v1 = 1, v2 = 2)
}
\arguments{
\item{theta}{- vector of probability for each component}

\item{psi}{- specific probability for each variable in each component}

\item{v1}{- first variable}

\item{v2}{- second variable}
}
\value{
Fisher - Fisher test statistics
}
\description{
Test the independence of two variables
}

