% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncompleteData.R
\docType{data}
\name{IncompleteData}
\alias{IncompleteData}
\title{This is an example for incomplete dataset}
\description{
IncompleteData is an array with parameters n = 10, p = 2, d = 2, m=10, k = 2 and missing rate is 0.25.
}
\author{
Chaojie Wang \email{wang910930@163.com}
}
\keyword{data}

