% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitialPsi.R
\name{InitialPsi}
\alias{InitialPsi}
\title{initial psi}
\usage{
InitialPsi(p, d, k)
}
\arguments{
\item{p}{- number of variables}

\item{d}{- a vector which denotes the number of categories for each variable. It could be distinct among variables.}

\item{k}{- number of components}
}
\value{
psi - a list in that each component has equal weight
}
\description{
This function creates a psi list in that each component has equal weight
}
