% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointTable.R
\name{JointTable}
\alias{JointTable}
\title{Generate joint distribution table}
\usage{
JointTable(p, d, alpha = 0.5)
}
\arguments{
\item{p}{- number of variables}

\item{d}{- number of categories}

\item{alpha}{- prior parameter}
}
\value{
JT - generated joint distribution table
}
\description{
This function is used to generate general joint distribution table
}

