% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointTableEst.R
\name{JointTableEst}
\alias{JointTableEst}
\title{Joint distribution table estimation
This function is used to esitmate joint distribution table based on theta and psi}
\usage{
JointTableEst(theta, psi)
}
\arguments{
\item{theta}{- vector of probability for each component}

\item{psi}{- specific probability for each variable in each component}
}
\value{
table - joint distribution table
}
\description{
Joint distribution table estimation
This function is used to esitmate joint distribution table based on theta and psi
}

