% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kIdentifier.R
\name{kIdentifier}
\alias{kIdentifier}
\title{Identify the suitable number of components k}
\usage{
kIdentifier(data, d, TT = 1000, alpha = 0.25)
}
\arguments{
\item{data}{- data in matrix formation with n rows and p columns}

\item{d}{- number of categories for each variable}

\item{TT}{- number of iterations in Gibbs sampler, default value is 1000. T should be an even number for 'burn-in'.}

\item{alpha}{- hyperparameter that could be regarded as the pseudo-count of the number of samples in the new component}
}
\value{
k_est - posterior estimation of k

k_track - track of k in the iteration process
}
\description{
This function is used to find the suitable number of components k.
}
\examples{
# dimension parameters
n<-200; p<-5; d<-rep(2,p);
# generate complete data
Complete<-GenerateData(n, p, d, k = 3)
# mask percentage of data at MCAR
Incomplete<-Complete
Incomplete[sample(1:n*p,0.2*n*p,replace = FALSE)]<-NA
# k identify
K<-kIdentifier(data = Incomplete, d, TT = 10)
}
