% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixToArray.R
\name{MatrixToArray}
\alias{MatrixToArray}
\title{Matrix To Array}
\usage{
MatrixToArray(data_m, d)
}
\arguments{
\item{data_m}{- a matrix format of observed dataset}

\item{d}{- number of choices}
}
\value{
data_a - an array format of observed dataset
}
\description{
If the number of trials m = 1, the observed dataset may be a matrix format. This function transforms observed dataset into array format, which is used in DPMCM package
}
\note{
n - number of observations

p - number of variables
}

