% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixToFactor.R
\name{MatrixToFactor}
\alias{MatrixToFactor}
\title{Matrix To Factor}
\usage{
MatrixToFactor(data_m, d)
}
\arguments{
\item{data_m}{- data in matrix formation}

\item{d}{- number of categoriesin each variable}
}
\value{
data_f - data in factor formation
}
\description{
This function is used to transform data from matrix into factor.
}

