% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdirichlet.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Estimate theta and psi in multinomial mixture model}
\usage{
rdirichlet(n = 1, alpha = c(1, 1))
}
\arguments{
\item{n}{- sample size}

\item{alpha}{- parameters in Dirichlet distribution}
}
\value{
out - generated data
}
\description{
This function is generate random sample from Dirichlet distribution
}
\examples{
# dimension parameters
rdirichlet(n=10,alpha=c(1,1,1))
}
