% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleJointTable.R
\name{SampleJointTable}
\alias{SampleJointTable}
\title{Sample joint distribution table}
\usage{
SampleJointTable(data, d)
}
\arguments{
\item{data}{- samples are observed}

\item{d}{- number of categories}
}
\value{
SJT - sample joint distribution table
}
\description{
This function is used to calcualte sample joint distribution table given samples.
}

