% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignTest.R
\name{SignTest}
\alias{SignTest}
\title{Sign test}
\usage{
SignTest(x, y, method = "One-tail")
}
\arguments{
\item{x}{- first variable}

\item{y}{- second variable}

\item{method}{- In "One-tail" method, the p-value is computed by Pr(x>y).
In "Two-tail" method, the p-value is computed by Pr(x>y)+Pr(x<y).}
}
\value{
p_value - p value in the test
}
\description{
Test for consistent differences between pairs of observations
}

