% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation.R
\name{Correlation}
\alias{Correlation}
\title{Correlation estimation}
\usage{
Correlation(theta, psi, v1 = 1, v2 = 2)
}
\arguments{
\item{theta}{- vector of probability for each component}

\item{psi}{- specific probability for each variable in each component}

\item{v1}{- first variable}

\item{v2}{- second variable}
}
\value{
Cor - correlation
}
\description{
Calculate Pearson's correlation between two specific variables
}

