% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{Imputation}
\alias{Imputation}
\title{Imputation}
\usage{
Imputation(data, theta, psi)
}
\arguments{
\item{data}{- incomplete dataset}

\item{theta}{- vector of probability for each component}

\item{psi}{- specific probability for each variable in each component}
}
\value{
ImputedData - dataset has been imputated.
}
\description{
This function is used to perform multiple imputation for missing data given the joint distribution.
}
