% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutualInfo.R
\name{MutualInfo}
\alias{MutualInfo}
\title{mutual information}
\usage{
MutualInfo(JT, v1, v2)
}
\arguments{
\item{JT}{- joint distribution table used}

\item{v1}{- first variable}

\item{v2}{- second variable}
}
\value{
MI - mutual information between two variables

MIC - maximal information coefficient between two variables
}
\description{
This function is used to calculate the mutual information (MI) and maximal information coefficient (MIC) between two variables.
}

