% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParEst.R
\name{ParEst}
\alias{ParEst}
\title{Estimate theta and psi in multinomial mixture model}
\usage{
ParEst(data, d, k, T = 1000)
}
\arguments{
\item{data}{- data in matrix formation with n rows and p columns}

\item{d}{- number of categories for each variable}

\item{k}{- number of components}

\item{T}{- number of iterations in Gibbs sampler, default value is 1000. T should be an even number for 'burn-in'.}
}
\value{
theta - vector of probability for each component

psi - specific probability for each variable in each component
}
\description{
This function is used to estimate theta and psi in multinomial mixture model given the number of components k.
}

